Deprecate 'gatherer.py' in favor of 'geographic_gatherer.py'
