+++ setup.py
@@ -50,6 +50,7 @@ setup(name="pytroll_collectors",
       scripts=['bin/trollstalker.py',
                'bin/trollstalker2.py',
                'bin/gatherer.py',
+               'bin/geographic_gatherer.py',
                'bin/segment_gatherer.py',
                'bin/cat.py',
                'bin/catter.py',
